pyproject.toml
[project]
name = "smartlearn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
    "sqlalchemy>=2.0.36",
    "aiomysql>=0.2.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.10.0",
    "email-validator>=2.2.0",
]
